+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 	* Reviving friends
 	* Removing thorns
 	* Opening Cursed Luggage
+	* Passing and Catching thrown items
 * Added a new config for Style Rank Multiplier, to change how many points are needed for each rank. Increase to make it take longer to rank up.
 * Added an internal trick cooldown to ensure that the same trick cannot be activated more than once every .1 seconds. Safety feature to avoid being spammed with tricks in glitched situations.